+++ packages/apiUtils/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @moralisweb3/api-utils
 
+## 2.7.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @moralisweb3/core@2.7.3
+
+## 2.7.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @moralisweb3/core@2.7.2
+
 ## 2.7.1
 
 ### Patch Changes