+++ packages/client/firebaseAuthUtils/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @moralisweb3/client-firebase-auth-utils
 
+## 2.7.3
+
+## 2.7.2
+
 ## 2.7.1
 
 ## 2.7.0