+++ packages/streams/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @moralisweb3/streams
 
+## 2.7.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @moralisweb3/core@2.7.3
+  - @moralisweb3/api-utils@2.7.3
+  - @moralisweb3/evm-utils@2.7.3
+
+## 2.7.2
+
+### Patch Changes
+
+- [#784](https://github.com/MoralisWeb3/Moralis-JS-SDK/pull/784) [`a1bbd61f`](https://github.com/MoralisWeb3/Moralis-JS-SDK/commit/a1bbd61f53b57feeeae1b3341f64d4293fba1497) Thanks [@sogunshola](https://github.com/sogunshola)! - Fix errors due to changes in webhook data for log parser function and update parse server streams docs
+
+- Updated dependencies []:
+  - @moralisweb3/core@2.7.2
+  - @moralisweb3/api-utils@2.7.2
+  - @moralisweb3/evm-utils@2.7.2
+
 ## 2.7.1
 
 ### Patch Changes